requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
aiohttp==3.9.1
schedule==1.2.0
flask==2.3.3
gunicorn==21.2.0
